# app/static/openapi.yaml
openapi: 3.0.1
servers:
- url: https://resources.operationcode.org/api/v1
  description: Default Server
info:
  title: Operation Code Resources API
  version: '1.0'
  x-logo:
    url: 'https://operation-code-assets.s3.us-east-2.amazonaws.com/branding/logos/small-blue-logo.png'
    altText: 'Operation Code Logo'
    backgroundColor: '#252e3e'
    href: 'https://operationcode.org/contact'
  description: |
    # Introduction
    Operation Code Resources API is a collection of programming and cyber security learning resources curated by and for the Operation Code community.

    # Authentication
    Create (POST) and update (PUT) routes are protected and require an API key. API key requests to the `/apikey` endpoint require valid Operation Code credentials in the request body.

    If you are not already a member of our community, join Operation Code [HERE](https://operationcode.org/join).

    When you create a request to an authenticated route, you *MUST* include the header `x-apikey` along with your API key.

    # Versioning
    To ensure contracts with the api are not broken with future updates, you may supply a version number in your request header.
    Include an `X-Api-Version` header with a valid version in your request:
      ```
      X-Api-Version: 1.0
      ```
    ### Default Versions
...


  contact:
    name: Operation Code
    url: https://operationcode.org/contact
paths:
...
  /languages:
    get:
      summary: List Languages
      description: Gets list of all available language entries.
      tags:
        - List Languages
      responses:
        200:
          description: A list of languages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
              example:
                apiVersion: '1.0'
                languages:
                  - id: 1
                    name: 'multiple'
                  - id: 2
                    name: 'Python'
                  - id: 3
                    name: 'Unix/Linux/*nix'
                  - id: 4
                    name: 'Bash'
                  - id: 5
                    name: 'C'
                  - id: 6
                    name: 'Sysadmin'
                  - id: 7
                    name: 'JavaScript'
                  - id: 8
                    name: 'HTML'
                  - id: 9
                    name: 'PHP'
                  - id: 10
                    name: 'Ruby'
                  - id: 11
                    name: 'Aspectwerkz'
                  - id: 12
                    name: 'Artificial Intelligence'
                  - id: 13
                    name: 'C++'
                  - id: 14
                    name: 'MongoDB'
                  - id: 15
                    name: 'SQL'
                  - id: 16
                    name: 'R'
                  - id: 17
                    name: 'GraphQL'
                  - id: 18
                    name: 'Swift'
                  - id: 19
                    name: 'Android'
                  - id: 20
                    name: 'CSS'
                  - id: 21
                    name: 'jQuery'
                  - id: 22
                    name: 'Git'
                status: 'ok'
                status_code: 200
                total_count: 22
        404:
          $ref: '#/components/responses/NotFound'
  /languages/{id}:
    get:
      summary: Get Language
      description: Gets more details about a specific language, given the language ID
      tags:
        - Get Language
      parameters:
      - name: id
        in: path
        description: Language ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: Language details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
              example:
                apiVersion: '1.0'
                language:
                  id: 2
                  name: 'Python'
                status: 'ok'
                status_code: 200
        404:
          $ref: '#/components/responses/NotFound'
  /resources/{id}:
    get:
      summary: Get Resource
      description: Gets more details about a specific resource, given the resource ID
      tags:
        - Get Resource
      parameters:
      - name: id
        in: path
        description: Resource ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: Resource details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
              example:
                apiVersion: '1.0'
                resource:
                  category: 'Books'
                  created_at: '2019-06-15 17:55:32'
                  downvotes: 0
                  id: 3
                  languages: []
                  last_updated: '2019-06-15 17:55:32'
                  name: 'Free Tech Books'
                  notes: 'Focuses on general computer science concepts rather than a specific language'
                  free: true
                  times_clicked: 0
                  upvotes: 0
                  url: 'http://www.freetechbooks.com/'
                  user_vote_direction: null
                status: 'ok'
                status_code: 200
        404:
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update Resource
      description: Updates an existing learning resource entry. The request body should be a JSON object containing the properties to be updated and their corresponding new values.
      tags:
        - Update Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceRequest'
            example:
              name: 'Updated Book Title'
        required: true
      parameters:
        - name: id
          in: path
          description: Resource ID
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
              example:
                apiVersion: '1.0'
                resource:
                  category: 'Tutorial'
                  created_at: '2019-07-21 15:41:19'
                  downvotes: 0
                  id: 2137
                  languages: []
                  last_updated: '2019-07-22 16:05:53'
                  name: 'Updated Book Title'
                  notes: null
                  free: true
                  times_clicked: 3
                  upvotes: 1
                  url: 'http://www.test.com'
                  user_vote_direction: null
                status: 'ok'
                status_code: 200
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/UnauthorizedNoApiKey'
        422:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                apiVersion: '1.0'
                errors:
                  invalid-params:
                      message: 'The following params were invalid: url. Resource id 2137 already has this URL.'
                      params:
                        - 'url'
                      resource: 'https://resources.operationcode.org/api/v1/2137'
                status: 'Unprocessable Entity'
                status_code: 422
      security:
      - ApiKeyAuth: []
  /resources/{id}/click:
    put:
      summary: Update Resource - Increment Number of Views
      description: |
        Increments the `times_clicked` property of a resource. The intention of this endpoint is to collect data about how popular a resource is based on how many times it is accessed. Please set up your integration to increment the `times_clicked` property when a user clicks through to view a resource.
      tags:
        - Update Resource
      parameters:
      - name: id
        in: path
        description: Resource ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
              example:
                apiVersion: '1.0'
                resource:
                    category: 'Books'
                    created_at: '2019-04-28 22:18:34'
                    downvotes: 0
                    id: 10
                    languages: []
                    last_updated: '2019-07-26 13:47:52'
                    name: 'Thinking Forth'
                    notes: null
                    free: true
                    times_clicked: 2
                    upvotes: 0
                    url: 'http://thinking-forth.sourceforge.net/'
                    user_vote_direction: null
                status: 'ok'
                status_code: 200
        404:
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
    Error:
      type: object
      properties:
        apiVersion:
          type: string
          description: API version
        errors:
          $ref: '#/components/schemas/ErrorData'
        status:
          type: string
          description: Status message
        status_code:
          type: integer
          description: Status code
    ErrorData:
      type: object
      description: The response errors
      properties:
        missing-params:
          type: object
          properties:
            params:
              description: The associated parameters.
              type: array
              items:
                type: string
            message:
              description: Error message
              type: string
        invalid-params:
          type: object
          properties:
            params:
              description: The associated parameters.
              type: array
              items:
                type: string
            message:
              description: Error message
              type: string
            resource:
              description: An href link to the resource with a duplicate url
              type: string
        unauthorized:
          type: object
          properties:
            message:
              description: Error message
              type: string
        missing-body:
          type: object
          properties:
            message:
              description: Error message
              type: string
        unprocessable-entity:
          type: object
          properties:
            message:
              description: Error message
              type: string
        bad-request:
          type: object
          properties:
            message:
              description: Error message
              type: string
        not-found:
          type: object
          properties:
            message:
              description: Error message
              type: string
...
    Resource:
      type: object
      properties:
        apiVersion:
          type: string
          description: API version
        resource:
          $ref: '#/components/schemas/ResourceData'
        status:
          type: string
          description: Status message
        status_code:
          type: integer
          description: Status code
...
    Language:
      type: object
      description: Language or topic presented in the resource
      properties:
        apiVersion:
          type: string
          description: API version
        languages:
          $ref: '#/components/schemas/LanguageData'
        status:
          type: string
          description: Status message
        status_code:
          type: integer
          description: Status code
        total_count:
          type: integer
          description: Total languages retrieved
    LanguageData:
      type: object
      properties:
        id:
          type: integer
          description: Language ID
        name:
          type: string
          description: Language Name
...
    ResourceRequest:
      type: object
      properties:
        category:
          type: string
          description: Category
        languages:
          type: array
          description: List of languages
          items:
            type: string
        name:
          type: string
          description: Resource name
        notes:
          type: string
          description: Resource notes
        free:
          type: boolean
          description: True or false for whether resource is free
        url:
          type: string
          description: Resource url
    ResourceData:
      type: object
      properties:
        category:
          type: string
          description: Category
        created_at:
          type: string
          format: date-time
          description: Date created
        downvotes:
          type: integer
          description: Negative ratings
        id:
          type: integer
          description: Resource ID
        last_updated:
          type: string
          format: date-time
          description: Date updated
        languages:
          type: array
          description: List of languages
          items:
            type: string
        name:
          type: string
          description: Resource name
        notes:
          type: string
          description: Resource notes
        free:
          type: boolean
          description: True or false for whether resource is free
        upvotes:
          type: integer
          description: Positive ratings
        url:
          type: string
          description: Resource url
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            apiVersion: '1.0'
            errors:
              not-found:
                message: 'The resource you requested does not exist.'
            status: 'Not Found'
            status_code: 404
...
    UnauthorizedNoApiKey:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            apiVersion: '1.0'
            errors:
              unauthorized:
                message: 'You must provide a valid API token in the x-apikey header.'
            status: 'Unauthorized'
            status_code: 401
...
  securitySchemes:
    ApiKeyAuth:
      description: API key authentication
      type: apiKey
      name: x-apikey
      in: header

##################
# Groups and Tags
##################
...